Call PRI callback on HUP
The code path handling the HUP did not call the pri callback. If only
PRI was being watched on a socket on which an ERR/HUP arrives, the
socket will be removed from the poll and always marked as processable
instead. As the pri callback is not called, the socket will not be
processed, causing the polling code to loop indefinitely.