Fixed blob size problem when the blob is swapped out
When a blob is growing and grow out the available memory it will swapped
out. But when this happen the allocated size of the blob should change
from the nearest two power to the exact size of the file.

This fix do this in a tricky way to change the alloc_size calculation to
dynamic. The only problem what could happen is when the blob is
swapped in in the time between the granted allocation and the real
allocation but this could not happen, because the blob should be locked in
the time of allocation.