src/zorp: add compat.h containing GLib-related compatibility macros
Recent GLib versions obsolete features and not always provide appropriate
macros so that dependant code can be compiled with both old and new versions
of GLib.

This header contains macros doing the right thing for the version of GLib
being used.

The header has been moved here from Zorp, because it turned out that not
only Zorp relies on these macros but we need them in ZCV, too.