src/streamssl: add possibility to store related handshake objects
The lifetime of SSL handshake objects should be tied to the stream the
handshake was done on instead of the proxy instance itself. This way the
handshake object(s) can be freed as soon as the SSL stream gets closed.

The idea is that none of our OpenSSL callbacks can be called once the
stream is closed so that's a safe point to free the related handshake
structure.

This patch adds the required infrastructure to store user data in an SSL
stream (together with its destroy method) so that the stream can be made
responsible of managing the lifetime of related handshake objects.